fix: move theme package into ui package
Move the theme package into the ui package to pass the corresponding
tests.
