+++ app/src/main/java/com/github/se/studentconnect/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.github.se.studentconnect.theme
+package com.github.se.studentconnect.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
